+++ packages/guide/src/app/journal/components/base-journal-page/base-journal-page.component.ts
@@ -1,31 +1,49 @@
 import { ChangeDetectorRef, Component, inject } from '@angular/core';
 import { Item, JournalOrder } from '@ci/data-types';
-import { combineLatest, map, Observable } from 'rxjs';
+import { combineLatest, map, Observable, startWith } from 'rxjs';
 import { DatabaseService } from '../../../shared/services/database.service';
 import { UiIcon } from '../../../shared/enums/ui-icon.enum';
 import { MediaMatcher } from '@angular/cdk/layout';
+import { Season } from '../../../shared/enums/season.enum';
+import { FormControl } from '@angular/forms';
+
+export interface BaseJournalPageComponent<D> {
+    filterPredicate?(foundEntry: D, filterValues: BaseJournalPageComponent<D>['formControl']['value']): boolean;
+}
 
 @Component({
     template: ''
 })
-export class BaseJournalPageComponent<D> {
+export class BaseJournalPageComponent<D extends ({ item: Item } | Item)> {
 
     uiIcon = UiIcon;
     tabs: { title: string; data: Observable<D[]> }[] = [];
     openDrawer = false;
     selectedEntity?: D;
 
+    season = Season;
+    formControl: FormControl<Season[]>;
+
     protected readonly _database: DatabaseService = inject(DatabaseService);
 
-    getFilteredJournalData<T extends { item: Item } | Item>(journalOrder$: Observable<JournalOrder[]>, data: Observable<T[]>): Observable<T[]> {
-        return combineLatest([journalOrder$, data]).pipe(map(
-            ([orders, entries]) => {
-                const res: T[] = [];
+    getFilteredJournalData(journalOrder$: Observable<JournalOrder[]>, data: Observable<D[]>): Observable<D[]> {
+        console.log(this.formControl.value);
+        return combineLatest([journalOrder$, data, this.formControl.valueChanges.pipe(startWith(this.formControl.value))]).pipe(map(
+            ([orders, entries, filterValues]) => {
+                console.log({orders, entries, filterValues});
+                const res: D[] = [];
+
 
                 orders.sort((a, b) => a.order > b.order ? 1 : -1).forEach(journalOrder => {
-                    const foundEntry: T | undefined = entries.find(f => 'item' in f ? (f.item.id === journalOrder.key) : (f.id === journalOrder.key));
+                    const foundEntry: D | undefined = entries.find(f => 'item' in f ? (f.item.id === journalOrder.key) : (f.id === journalOrder.key));
                     if (foundEntry) {
-                        res.push(foundEntry);
+                        if (!this.filterPredicate)
+                            res.push(foundEntry);
+                        if (this.filterPredicate && this.filterPredicate(foundEntry, filterValues)) {
+                            res.push(foundEntry);
+                        }
+
+
                     }
                 });
                 return res;
@@ -43,6 +61,7 @@ export class BaseJournalPageComponent<D> {
         this.mobileQuery = this.media.matchMedia('(max-width: 600px)');
         this._mobileQueryListener = () => this.changeDetectorRef.detectChanges();
         this.mobileQuery.addListener(this._mobileQueryListener);
+        this.formControl = new FormControl<Season[]>([Season.SPRING, Season.SUMMER, Season.FALL, Season.WINTER], {nonNullable: true});
     }
 
     showDetails(fishEntry?: D) {