+++ packages/guide/src/app/journal/components/caught/caught.component.ts
@@ -1,6 +1,8 @@
 import { Component } from '@angular/core';
 import { Critter, Fish } from '@ci/data-types';
 import { BaseJournalPageComponent } from '../base-journal-page/base-journal-page.component';
+import { getTruthyValues } from '@ci/util';
+import { Season } from '../../../shared/enums/season.enum';
 
 @Component({
     selector: 'app-caught',
@@ -8,6 +10,7 @@ import { BaseJournalPageComponent } from '../base-journal-page/base-journal-page
     styleUrls: ['./caught.component.scss'],
 })
 export class CaughtComponent extends BaseJournalPageComponent<Fish | Critter> {
+    showTable = false;
 
     constructor() {
         super();
@@ -35,6 +38,23 @@ export class CaughtComponent extends BaseJournalPageComponent<Fish | Critter> {
             },
         ];
 
+    }
+
+    private _isFish(array: (Critter | Fish) | undefined): array is Fish {
+        return !!array && 'fishName' in array;
+    }
+
+
+    override filterPredicate(foundEntry: Fish | Critter, filterValues: Season[]): boolean {
+
+        if (filterValues.length === 0) return false;
+
+        if (this._isFish(foundEntry)) {
+            const seasonString = getTruthyValues(foundEntry.spawnSeason).toLowerCase();
+            return seasonString === 'any' || filterValues.some(season => seasonString.includes(('' + season).toLowerCase()));
+        }
+
+        return true;
     }