+++ packages/guide/src/app/journal/components/tables/caught-table/caught-table.component.spec.ts
@@ -1,0 +1,22 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { CaughtTableComponent } from './caught-table.component';
+
+describe('FishTableComponent', () => {
+  let component: CaughtTableComponent;
+  let fixture: ComponentFixture<CaughtTableComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [CaughtTableComponent],
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(CaughtTableComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});