+++ packages/guide/src/app/journal/components/tables/found-table/found-table.component.ts
@@ -1,0 +1,14 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-found-table',
+  templateUrl: './found-table.component.html',
+  styleUrls: ['./found-table.component.scss'],
+})
+export class FoundTableComponent implements OnInit {
+  constructor() {
+  }
+
+  ngOnInit(): void {
+  }
+}