+++ packages/guide/src/app/journal/components/tables/produce-table/produce-table.component.ts
@@ -1,0 +1,14 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-produce-table',
+  templateUrl: './produce-table.component.html',
+  styleUrls: ['./produce-table.component.scss'],
+})
+export class ProduceTableComponent implements OnInit {
+  constructor() {
+  }
+
+  ngOnInit(): void {
+  }
+}